Resolve cherry-pick conflicts against the 25.8 base:

- CompressionCodecMultiple.cpp: keep the two added bounds checks.
- CompressionCodecT64.cpp: keep the 25.8 version unchanged. 25.8 has no
  unaligned trailing-byte memcpy path and already rejects a compressed
  size smaller than the min/max header, so neither T64 overflow is
  reachable here.
- gtest_compressionCodec.cpp: add the four regression tests on top of the
  25.8 test file; the two T64 cases assert 25.8 rejects short input at
  its existing header-size guard.

+++ src/IO/examples/read_float_perf.cpp
@@ -1,8 +1,3 @@
-// NOLINTBEGIN(clang-analyzer-core.UndefinedBinaryOperatorResult,clang-analyzer-optin.core.EnumCastOutOfRange)
-// False positives: clang-analyzer traces into fast_float library headers and reports
-// UndefinedBinaryOperatorResult in ascii_number.h and EnumCastOutOfRange in float_common.h.
-// These cannot be suppressed via NOLINT at call sites because the diagnostics are reported
-// at the library header locations, not at the call site.
 #include <string>
 
 #include <iostream>
@@ -60,7 +55,7 @@ void NO_INLINE loop(ReadBuffer & in, WriteBuffer & out)
 
     watch.stop();
     out << "Read in " << watch.elapsedSeconds() << " sec, "
-        << formatReadableSizeWithBinarySuffix(static_cast<double>(in.count()) / watch.elapsedSeconds()) << "/sec, result = " << sum << "\n";
+        << formatReadableSizeWithBinarySuffix(in.count() / watch.elapsedSeconds()) << "/sec, result = " << sum << "\n";
 }
 
 
@@ -89,4 +84,3 @@ catch (const Exception & e)
     std::cerr << e.what() << ", " << e.displayText() << std::endl;
     return 1;
 }
-// NOLINTEND(clang-analyzer-core.UndefinedBinaryOperatorResult,clang-analyzer-optin.core.EnumCastOutOfRange)
